third_party/WebKit/LayoutTests/http/tests/websocket/invalid-subprotocols.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("WebSocket should fail if invalid subprotocols are specified");

var url = "ws://127.0.0.1:8880/websocket/tests/simple";
shouldThrow("new WebSocket(url, 'a b')", '"SyntaxError: Failed to construct \'WebSocket\': The subprotocol \'a b\' is invalid."');
shouldThrow("new WebSocket(url, 'a,b')", '"SyntaxError: Failed to construct \'WebSocket\': The subprotocol \'a,b\' is invalid."');
shouldThrow("new WebSocket(url, 'a\\0b')", '"SyntaxError: Failed to construct \'WebSocket\': The subprotocol \'a\\\\u0000b\' is invalid."');
shouldThrow("new WebSocket(url, '\\u3042')", '"SyntaxError: Failed to construct \'WebSocket\': The subprotocol \'\\\\u3042\' is invalid."');
shouldThrow("new WebSocket(url, ['a', 'b', 'c d'])", '"SyntaxError: Failed to construct \'WebSocket\': The subprotocol \'c d\' is invalid."');
shouldThrow("new WebSocket(url, ['a', 'b', 'c', 'a'])", '"SyntaxError: Failed to construct \'WebSocket\': The subprotocol \'a\' is duplicated."');

</script>
</body>
</html>
